Adding GA property ID
This commit also adds a development config file that clears the GA
property ID so GA hits are not sent while developing. The Rakefile
is updated to apply this hierarchy of config files, so when `rake`
is used to run the site, the config files are applied correctly.

Part of #1.