Remove defaults
Since updating dependencies in 62c4891, the Atom feed plugin
(jekyll-feed) is failing with

> Cannot sort a null object. in feed.xml

It ends up that the list of posts being provided to the feed plugin
was empty. The defaults removed by this commit was somehow the
culprit. Those defaults were added in the past to make the tagging
system work. As far as I can tell now, the tagging system works
without the defaults.